+++ lib/shared/themes/default.ts
@@ -7,7 +7,7 @@ const theme: ThemeProps = {
 	backgroundColor: Color.white,
 	borderColor: Color.silver,
 	buttonColor: Color.gray,
-	chevronColor: Color.gray,
+	chevronColor: Color.lightgray,
 	color: Color.black,
 	fontFamily: "OpenSans, Georgia, serif",
 	headerForegroundColor: Color.white,