+++ lib/tabs/TabContainer.test.tsx
@@ -44,7 +44,7 @@ test("Test TabContainer retrieval functions", () => {
 	const container = ctl.instance() as TabContainer;
 	expect(container).toBeDefined();
 
-	const id: string = container.tabs.get(0).props["id"];
+	const id: string = container.tabs[0].props["id"];
 	let tab;
 	let idx;
 
@@ -156,7 +156,7 @@ test("Test selecting the first tab within the TabContainer", () => {
 	const container = ctl.instance() as TabContainer;
 	expect(container).toBeDefined();
 
-	expect(container.tabs.size).toBe(3);
+	expect(container.tabs.length).toBe(3);
 	const firstTab = ctl
 		.find(".ui-tab")
 		.first()
@@ -164,25 +164,28 @@ test("Test selecting the first tab within the TabContainer", () => {
 		.first();
 	firstTab.simulate("click");
 	expect(select).toHaveBeenCalled();
-	expect(container.tabs.size).toBe(3);
+	expect(container.tabs.length).toBe(3);
 });
 
 test("Test removing the first item from the TabContainer", () => {
 	const remove = jest.fn();
+	const closeButton1 = jest.fn();
+	const closeButton3 = jest.fn();
 	const ctl = mount(
 		<TabContainer onRemove={remove}>
-			<Tab title='tab #1'>#1</Tab>
+			<Tab onClose={closeButton1} title='tab #1'>
+				#1
+			</Tab>
 			<Tab title='tab #2'>#2</Tab>
-			<Tab title='tab #3'>#3</Tab>
+			<Tab onClose={closeButton3} title='tab #3'>
+				#3
+			</Tab>
 		</TabContainer>
 	);
 
 	assert(ctl);
 	expect(ctl).toMatchSnapshot();
-	const container = ctl.instance() as TabContainer;
-	expect(container).toBeDefined();
 
-	expect(container.tabs.size).toBe(3);
 	const firstTab = ctl
 		.find(".ui-tab")
 		.first()
@@ -190,7 +193,7 @@ test("Test removing the first item from the TabContainer", () => {
 		.first();
 	firstTab.simulate("click");
 	expect(remove).toHaveBeenCalled();
-	expect(container.tabs.size).toBe(2);
+	expect(closeButton1).toHaveBeenCalled();
 
 	const lastTab = ctl
 		.find(".ui-tab")
@@ -199,7 +202,7 @@ test("Test removing the first item from the TabContainer", () => {
 		.first();
 	lastTab.simulate("click");
 	expect(remove).toHaveBeenCalledTimes(2);
-	expect(container.tabs.size).toBe(1);
+	expect(closeButton3).toHaveBeenCalled();
 });
 
 test("Test the TabContainer with the previous and next buttons", () => {
@@ -218,7 +221,7 @@ test("Test the TabContainer with the previous and next buttons", () => {
 	expect(container).toBeDefined();
 
 	// Assert that the current tab is the first in the list
-	expect(ctl.state("selectedTab")).toBe(container.tabs.get(0).props["id"]);
+	expect(ctl.state("selectedTab")).toBe(container.tabs[0].props["id"]);
 
 	const prevButton = ctl
 		.find(".ui-tab-navigation")
@@ -234,10 +237,10 @@ test("Test the TabContainer with the previous and next buttons", () => {
 	// Click the "next" button and move from 1 -> 2
 	nextButton.simulate("click");
 	expect(select).toHaveBeenCalled();
-	expect(ctl.state("selectedTab")).toBe(container.tabs.get(1).props["id"]);
+	expect(ctl.state("selectedTab")).toBe(container.tabs[1].props["id"]);
 
 	// Click the "previous" button and move from 2 -> 1
 	prevButton.simulate("click");
 	expect(select).toHaveBeenCalledTimes(2);
-	expect(ctl.state("selectedTab")).toBe(container.tabs.get(0).props["id"]);
+	expect(ctl.state("selectedTab")).toBe(container.tabs[0].props["id"]);
 });