+++ lib/tabs/TabContainer.tsx
@@ -16,11 +16,16 @@
  *      location={Location.bottom}
  *      nonavigation
  *      onRemove={(tab: any) => {
- *          debug(removing %o (id=${tab.props['id']})`, tab);
+ *          debug(removing %o (id=${tab.props['id']})`, tab.props["title"]);
  *      }}
  *      onSelect={(tab: any, previous: any) => {
- *          debug(`new: %o (id=${tab.props['id']}),
- *              old: %o (id=${previous.props['id']})`, tab, previous);
+ *          debug(
+ *              `selected: %o (id=${tab.props["id"]}), previous: %o (id=${
+ *                  previous.props["id"]
+ *              })`,
+ *              tab.props["title"],
+ *              previous.props["title"]
+ *          );
  *      }}
  * >
  *     <Tab title="tab #1">#1<br/><br/>{randomText}</Tab>
@@ -70,11 +75,12 @@
  * @module TabContainer
  */
 
-const debug = require("debug")("TabContainer");
+// const debug = require("debug")("TabContainer");
 
 import autobind from "autobind-decorator";
-import {List} from "immutable";
+import _ from "lodash";
 import * as React from "react";
+import {unitToNumber} from "util.calc";
 import {Keys} from "util.keys";
 import {nilEvent} from "util.toolbox";
 import {Button} from "../button";
@@ -101,17 +107,23 @@ export interface TabContainerProps extends BaseProps {
 }
 
 export function getDefaultTabContainerProps(): TabContainerProps {
-	return {
-		...getDefaultBaseProps(),
-		location: Location.top,
-		maxTabs: 5,
-		noclose: false,
-		nonavigation: false,
-		obj: "TabContainer",
-		onRemove: nilEvent,
-		onSelect: nilEvent,
-		tabWidth: 120
-	};
+	return _.omitBy(
+		{
+			...getDefaultBaseProps(),
+			location: Location.top,
+			maxTabs: 5,
+			noborder: false,
+			noclose: false,
+			nonavigation: false,
+			obj: "TabContainer",
+			onRemove: nilEvent,
+			onSelect: nilEvent,
+			tabWidth: 100,
+			minHeight: "100px",
+			minWidth: "350px"
+		},
+		_.isNil
+	);
 }
 
 export interface TabContainerState extends BaseState {
@@ -126,7 +138,7 @@ export function getDefaultTabContainerState(): TabContainerState {
 }
 
 const TabBarHorizontal: any = css`
-	display: block;
+	display: flex;
 `;
 
 const TabBarVertical: any = css`
@@ -135,35 +147,87 @@ const TabBarVertical: any = css`
 `;
 
 const TabBarView: any = styled.div`
-	display: flex;
-	${(props: TabContainerProps) => props.xcss}
+	${(props: TabContainerProps) => {
+		if (
+			props.location === Location.top ||
+			props.location === Location.bottom
+		) {
+			return TabBarHorizontal;
+		} else {
+			return TabBarVertical;
+		}
+	}}
 `;
 
 const TabContainerView: any = styled.div`
+	display: flex;
+	height: 100%;
+	min-height: ${(props: TabContainerProps) => props.minHeight};
+	min-width: ${(props: TabContainerProps) => props.minWidth};
+	width: 100%;
+
 	${(props: TabContainerProps) =>
 		props.location === Location.top || props.location === Location.bottom
-			? ""
-			: "display: flex; flex-wrap: nowrap;"}
+			? "flex-direction: column;"
+			: "flex-wrap: nowrap;"}
 `;
 
 const TabContentHorizontal: any = css`
-	border: solid 1px ${(props: TabContainerProps) => props.theme.borderColor};
-	box-sizing: border-box;
-	min-height: 8em;
-	padding: 0 3px;
+	display: flex;
+
+	${(props: TabContainerProps) => {
+		if (props.location === Location.top) {
+			return "border-top: solid 1px " + props.theme.borderColor + ";";
+		} else if (props.location === Location.bottom) {
+			return "border-bottom: solid 1px " + props.theme.borderColor + ";";
+		} else {
+			return "";
+		}
+	}}
 `;
 
 const TabContentVertical: any = css`
-	border: solid 1px ${(props: TabContainerProps) => props.theme.borderColor};
-	box-sizing: border-box;
 	display: inline-flex;
-	min-height: 8em;
-	padding: 0 3px;
 	flex-grow: 1;
+
+	${(props: TabContainerProps) => {
+		if (props.location === Location.left) {
+			return "border-left: solid 1px " + props.theme.borderColor + ";";
+		} else if (props.location === Location.right) {
+			return "border-right: solid 1px " + props.theme.borderColor + ";";
+		} else {
+			return "";
+		}
+	}}
 `;
 
 const TabContentView: any = styled.div`
-	${(props: TabContainerProps) => props.xcss || ""}
+	${(props: TabContainerProps) => {
+		if (!props.noborder) {
+			return "border: solid 1px " + props.theme.borderColor + ";";
+		} else {
+			return "";
+		}
+	}}
+
+	box-sizing: border-box;
+	flex: 1;
+	padding: 0 3px;
+
+	${(props: TabContainerProps) => {
+		if (
+			props.location === Location.top ||
+			props.location === Location.bottom
+		) {
+			return TabContentHorizontal;
+		} else {
+			return TabContentVertical;
+		}
+	}}
+`;
+
+const TabNavigationTop: any = css`
+	margin-left: auto;
 `;
 
 const TabNavigationView: any = styled.div`
@@ -171,60 +235,28 @@ const TabNavigationView: any = styled.div`
 	display: flex;
 	float: right;
 
+	${(props: TabContainerProps) => {
+		switch (props.location) {
+			case Location.top:
+			case Location.bottom:
+				return TabNavigationTop;
+		}
+	}}
+
 	> .ui-button {
 		display: inline-block;
 		flex: unset;
 	}
 `;
 
-export const TabBar = (props: any) => (
-	<TabBarView
-		className='ui-tab-bar'
-		style={props.style}
-		xcss={
-			props.location === Location.top ||
-			props.location === Location.bottom
-				? TabBarHorizontal
-				: TabBarVertical
-		}
-	>
-		{props.tabs}
-		{props.navigation}
-	</TabBarView>
-);
-
-export const TabContent = (props: any) => (
-	<TabContentView
-		className='ui-tab-content'
-		style={props.style}
-		xcss={
-			props.location === Location.top ||
-			props.location === Location.bottom
-				? TabContentHorizontal
-				: TabContentVertical
-		}
-	>
-		{props.content}
-	</TabContentView>
-);
-
-export const TabNavigation = (props: any) =>
-	!props.nonavigation &&
-	!props.disabled && (
-		<TabNavigationView className='ui-tab-navigation'>
-			<Button iconName='chevron-left' onClick={props.handleLeftClick} />
-			<Button iconName='chevron-right' onClick={props.handleRightClick} />
-		</TabNavigationView>
-	);
-
 export class TabContainer extends BaseComponent<
 	TabContainerProps,
 	TabContainerState
 > {
-	private _containerWidth: number = 0;
 	private _keys: Keys;
 	private _tabContent: any = null;
-	private _tabs: any = List();
+	private _tabs: any = [];
+	private _removedTabs: any = [];
 
 	public static defaultProps: TabContainerProps = getDefaultTabContainerProps();
 
@@ -236,39 +268,12 @@ export class TabContainer extends BaseComponent<
 		// Initialize all of the tabs given to the container.  This will assign
 		// the id/key and remove any child that is not a Tab.
 		if (props.children) {
-			const children = React.Children.toArray(this.props.children);
-			let pos: number = 0;
-
-			for (const child of children) {
-				if (
-					child["type"] === Tab &&
-					(child["props"]["visible"] && !child["props"]["disabled"])
-				) {
-					this._tabs = this._tabs.set(
-						pos,
-						React.cloneElement(child as any, {
-							href: {
-								hiddenTabHandler: this.hiddenTabHandler,
-								selectHandler: this.selectHandler
-							},
-							id: this._keys.at(pos),
-							key: this._keys.at(pos)
-						})
-					);
-
-					pos++;
-				}
-			}
-
-			if (props.maxTabs > 0) {
-				this._tabs = this._tabs.slice(0, this.props.maxTabs);
-			}
+			this._tabs = this.updateChildren(this.props);
 		}
 
 		this.state = {
 			...getDefaultTabContainerState(),
-			selectedTab:
-				this._tabs.size > 0 ? this._tabs.get(0).props["id"] : null
+			selectedTab: this._tabs.length > 0 ? this._tabs[0].props["id"] : ""
 		};
 	}
 
@@ -301,8 +306,8 @@ export class TabContainer extends BaseComponent<
 	private handlePreviousTab() {
 		const idx: number = this.currentIdx;
 
-		if (this._tabs.size && idx - 1 >= 0) {
-			this.selectHandler(this._tabs.get(idx - 1));
+		if (this._tabs.length && idx - 1 >= 0) {
+			this.selectHandler(this._tabs[idx - 1]);
 		}
 	}
 
@@ -310,8 +315,8 @@ export class TabContainer extends BaseComponent<
 	private handleNextTab() {
 		const idx: number = this.currentIdx;
 
-		if (this._tabs.size && idx + 1 <= this._tabs.size - 1) {
-			this.selectHandler(this._tabs.get(idx + 1));
+		if (this._tabs.length && idx + 1 <= this._tabs.length - 1) {
+			this.selectHandler(this._tabs[idx + 1]);
 		}
 	}
 
@@ -325,27 +330,24 @@ export class TabContainer extends BaseComponent<
 		let id: string = null;
 
 		if (idx - 1 >= 0) {
-			id = this._tabs.get(idx - 1).props["id"];
-		} else if (idx + 1 <= this._tabs.size - 1) {
-			id = this._tabs.get(idx + 1).props["id"];
+			id = this._tabs[idx - 1].props["id"];
+		} else if (idx + 1 <= this._tabs.length - 1) {
+			id = this._tabs[idx + 1].props["id"];
 		}
 
-		this._tabs = this._tabs.splice(idx, 1);
 		this.selectedTab = id;
+		this._removedTabs.push(tab.props["id"]);
 		this.props.onRemove(tab);
 	}
 
 	@autobind
 	private selectHandler(tab: Tab) {
-		const [previous, idx] = this._getTab(this.state.selectedTab);
+		if (this.state.selectedTab) {
+			const previous = this._getTab(this.state.selectedTab)[0];
 
-		this.selectedTab = tab.props["id"];
-		this.props.onSelect(tab, previous ? previous : tab);
-		debug(
-			`selected tab: ${tab.props["id"]}, tab: %O, previous @ ${idx}: %O`,
-			tab,
-			previous
-		);
+			this.selectedTab = tab.props["id"];
+			this.props.onSelect(tab, previous ? previous : tab);
+		}
 	}
 
 	/**
@@ -361,7 +363,7 @@ export class TabContainer extends BaseComponent<
 		if (id) {
 			for (const [idx, tab] of this._tabs.entries()) {
 				if (tab.props["id"] === id) {
-					return [this._tabs.get(idx), idx];
+					return [this._tabs[idx], idx];
 				}
 			}
 		}
@@ -389,6 +391,55 @@ export class TabContainer extends BaseComponent<
 		return -1;
 	}
 
+	/**
+	 * The child elements under the container must be modified to add a select
+	 * handler and unique id values.  The id value for the tab is stable based
+	 * on the Keys class.
+	 * @param props {TabContainerProps} - the properties passed to the tab
+	 * container.
+	 * @return an array of Tabs components
+	 */
+	private updateChildren(props: TabContainerProps) {
+		let children = React.Children.map(
+			props.children,
+			(child: any, pos: number) => {
+				const id = this._keys.at(pos);
+				const selected =
+					id &&
+					this.state &&
+					"selectedTab" in this.state &&
+					this.state.selectedTab === id;
+
+				if (selected && !props.disabled) {
+					this._tabContent = child["props"].children;
+				}
+
+				const newChild = React.cloneElement(child, {
+					disabled: props.disabled,
+					href: {
+						hiddenTabHandler: this.hiddenTabHandler,
+						selectHandler: this.selectHandler
+					},
+					id,
+					key: id,
+					noclose: props.noclose,
+					orientation: props.location,
+					sizing: props.sizing,
+					selected: selected,
+					width: `${props.tabWidth}px`
+				});
+
+				return newChild;
+			}
+		).filter((child) => !this._removedTabs.includes(child.props.id));
+
+		if (props.maxTabs > 0) {
+			children = children.slice(0, this.props.maxTabs);
+		}
+
+		return children;
+	}
+
 	public render() {
 		this.updateClassName();
 
@@ -398,38 +449,8 @@ export class TabContainer extends BaseComponent<
 		// Removes non standard property to remove jest warning in test
 		const {onRemove, ...props} = this.props;
 
-		if (this._tabs.size > 0) {
-			for (const [idx, child] of this._tabs.entries()) {
-				const selected = this.state.selectedTab === child.props["id"];
-
-				if (selected && !props.disabled) {
-					this._tabContent = child["props"].children;
-				}
-
-				this._tabs = this._tabs.set(
-					idx,
-					React.cloneElement(child as any, {
-						disabled: props.disabled,
-						orientation: props.location,
-						sizing: props.sizing,
-						selected: selected,
-						width: `${props.tabWidth}px`
-					})
-				);
-			}
-
-			// Sets the default width of the content container for the component
-			// It uses the number of tabs and the current tab width to compute
-			// a reasonable size.
-
-			if (
-				props.location === Location.top ||
-				props.location === Location.bottom
-			) {
-				this._containerWidth = (this._tabs.size + 1) * props.tabWidth;
-			} else {
-				this._containerWidth = 5 * props.tabWidth;
-			}
+		if (this._tabs.length > 0) {
+			this._tabs = this.updateChildren(props);
 		} else {
 			// No more tabs, suppress content
 			this._tabContent = null;
@@ -442,25 +463,52 @@ export class TabContainer extends BaseComponent<
 			style["width"] = `${props.tabWidth}px`;
 		}
 
-		const tabNavigation = (
-			<TabNavigation
-				{...props}
-				handleLeftClick={this.handlePreviousTab}
-				handleRightClick={this.handleNextTab}
-				nonavigation={props.nonavigation}
-			/>
-		);
+		let tabNavigation: any = null;
+		if (!props.nonavigation) {
+			tabNavigation = (
+				<TabNavigationView {...props} className='ui-tab-navigation'>
+					<Button
+						color={this.theme.chevronColor}
+						iconName='chevron-left'
+						onClick={this.handlePreviousTab}
+					/>
+					<Button
+						color={this.theme.chevronColor}
+						iconName='chevron-right'
+						onClick={this.handleNextTab}
+					/>
+				</TabNavigationView>
+			);
+		}
 
 		const tabBar = (
-			<TabBar
+			<TabBarView
 				{...props}
-				navigation={tabNavigation}
-				style={style}
-				tabs={this._tabs}
-			/>
+				className='ui-tab-bar'
+				location={props.location}
+			>
+				{this._tabs}
+				{tabNavigation}
+			</TabBarView>
 		);
 
-		const tabContent = <TabContent {...props} content={this._tabContent} />;
+		const tabContent = (
+			<TabContentView
+				className='ui-tab-content'
+				location={props.location}
+				noborder={props.noborder}
+			>
+				{this._tabContent}
+			</TabContentView>
+		);
+
+		// The number +2 below represents an extra tab component and the
+		// navigation buttons for sizing purposes.
+		const sizingFactor = this._tabs.length + 2;
+		const minHeight =
+			unitToNumber(BaseComponent.lineHeightPX(this.props.sizing)) *
+			sizingFactor;
+		const minWidth = this.props.tabWidth * sizingFactor;
 
 		if (
 			props.location === Location.top ||
@@ -470,9 +518,10 @@ export class TabContainer extends BaseComponent<
 				<TabContainerView
 					{...props}
 					className={this.className}
-					style={{minWidth: `${this._containerWidth}px`}}
+					minHeight={`${minHeight}px`}
+					minWidth={`${minWidth}px`}
 				>
-					{this._tabs.size > 0 && tabBar}
+					{tabBar}
 					{tabContent}
 				</TabContainerView>
 			);
@@ -481,10 +530,11 @@ export class TabContainer extends BaseComponent<
 				<TabContainerView
 					{...props}
 					className={this.className}
-					style={{minWidth: `${this._containerWidth}px`}}
+					minHeight={`${minHeight}px`}
+					minWidth={`${minWidth}px`}
 				>
 					{tabContent}
-					{this._tabs.size > 0 && tabBar}
+					{tabBar}
 				</TabContainerView>
 			);
 		}